+++ src/mini_strace.c
@@ -425,6 +425,8 @@ static void run_tracer(pid_t child, int follow_forks, int show_timing, int summa
                     unsigned char sigprocmask_how_mask = sigprocmask_how_arg_mask(name);
                     unsigned char access_mode_mask = access_mode_arg_mask(name);
                     unsigned char clockid_mask = clockid_arg_mask(name);
+                    unsigned char clone_flags_mask = clone_flags_arg_mask(name);
+                    unsigned char clone3_args_mask = clone3_args_arg_mask(name);
                     unsigned char fd_mask = show_fd_paths ? fd_arg_mask(name) : 0;
                     const buffer_arg_entry *buf_entry = buffer_arg_lookup(name);
                     const buffer_arg_entry *sockaddr_entry = sockaddr_arg_lookup(name);
@@ -457,6 +459,10 @@ static void run_tracer(pid_t child, int follow_forks, int show_timing, int summa
                             format_access_mode(raw_args[i], argbuf[i], sizeof(argbuf[i]));
                         else if (clockid_mask & (1 << i))
                             format_clockid(raw_args[i], argbuf[i], sizeof(argbuf[i]));
+                        else if (clone_flags_mask & (1 << i))
+                            format_clone_flags(raw_args[i], argbuf[i], sizeof(argbuf[i]));
+                        else if (clone3_args_mask & (1 << i))
+                            format_clone3_flags(wpid, raw_args[i], argbuf[i], sizeof(argbuf[i]));
                         else if (buf_entry != NULL && i == buf_entry->buf_idx)
                             read_child_buffer(wpid, raw_args[i], raw_args[buf_entry->len_idx],
                                                argbuf[i], sizeof(argbuf[i]));